+++ app/(dashboard)/subjects/page.tsx
@@ -1,9 +1,11 @@
-import { DatabaseService } from "@/services/db-service";
 import { SubjectsList } from "./subjects-list";
+import { UserService } from "@/services/user-service";
+import { DatabaseService } from "@/services/db-service";
 
 export default async function SubjectsPage() {
-  const currentUser = (await DatabaseService.getCurrentUser())!;
-  const subjects = await DatabaseService.getSubjectsWithTeachers(currentUser.school_id);
+  const user = await DatabaseService.getCurrentUser();
+  const currentUser = await UserService.checkCurrentUser(user, "/subjects");
+  const subjects = await DatabaseService.getSubjectsWithTeachers(currentUser.active_school_id);
 
   return <SubjectsList subjects={subjects} />;
 }