+++ components/auth-button.tsx
@@ -1,19 +1,14 @@
 import Link from "next/link";
 import { Button } from "./ui/button";
-import { createClient } from "@/lib/supabase/server";
 import { LogoutButton } from "./logout-button";
+import { DatabaseService } from "@/services/db-service";
 
 export async function AuthButton() {
-  const supabase = await createClient();
-
-  // You can also use getUser() which will be slower.
-  const { data } = await supabase.auth.getClaims();
-
-  const user = data?.claims;
+  const user = await DatabaseService.getCurrentUser();
 
   return user ? (
     <div className="flex items-center gap-4">
-      Hey, {user.email}!
+      Hey, {user.full_name}!
       <LogoutButton />
     </div>
   ) : (