+++ components/navigation-sidebar/actions.ts
@@ -1,0 +1,7 @@
+"use server";
+
+import { DatabaseService } from "@/services/db-service";
+
+export async function updateUserActiveSchool(newSchoolId: string) {
+  return await DatabaseService.updateUserActiveSchool(newSchoolId);
+}